Move tempgres connection setup in tests to own class
